refactor(agent): extract run recovery phase
