+++ runtime/src/agent-pool/run-agent-orchestrator.ts
@@ -15,21 +15,15 @@ import { deleteSshConfig } from "../db.js";
 import { clearLiveSshConfig } from "../extensions/ssh-core.js";
 
 import {
-  decideAutomaticRecovery,
   getAutomaticRecoveryConfig,
-  getAutomaticRecoveryDelayMs,
-  isContextPressureFailure,
   isLengthStopFailure,
-  type RecoveryAttemptSnapshot,
-  type RecoveryClassifier,
-  type RecoveryStrategy,
 } from "./automatic-recovery.js";
-import { getAgentRuntimeConfig, getRecoveryPolicyConfig, getSessionStorageConfig, getToolUseBudget } from "../core/config.js";
+import { getAgentRuntimeConfig, getSessionStorageConfig, getToolUseBudget } from "../core/config.js";
 import { detectChannel } from "../router.js";
 import { pruneOrphanToolResults } from "./orphan-tool-results.js";
 import { writeAgentLog } from "./logging.js";
 import { createLogger, debugSuppressedError } from "../utils/logger.js";
-import { getSessionFileLineCount, getSessionFileSize, isRotationFallbackCompactionError, rotateSession } from "../session-rotation.js";
+import { getSessionFileLineCount, getSessionFileSize, rotateSession } from "../session-rotation.js";
 import { getCompactionSuccessCount, resetCompactionSuccessCount } from "./compaction.js";
 import { withChatContext } from "../core/chat-context.js";
 import {
@@ -39,18 +33,13 @@ import {
 } from "./prompt-utils.js";
 import {
   cancelScheduledIdleAutoCompaction,
-  buildFreshContextUsageUpdateEvent,
   clearCompactionFailureBackoff,
   getAutoCompactionTokenStatusForSession,
-  getCompactionContextReport,
   isCompactionCancellationError,
   maybeAutoCompactSessionAfterTurn,
   maybeAutoCompactSessionBeforePrompt,
-  noteCompactionFailure,
   noteCompactionSuccess,
-  runCompactionWithTimeout,
 } from "./compaction.js";
-import { buildPiclawCompactionEventFields, type PiclawCompactionTriggerMetadata } from "./compaction-trigger-context.js";
 import {
   inspectBlankTurnSessionDelta,
   isBlankTurnSessionDelta,
@@ -59,10 +48,13 @@ import {
 import {
   didPromptAdvanceSession,
   getSessionLeafId,
-  RECOVERY_CONTINUATION_PROMPT,
 } from "./context-pressure-retry.js";
+import {
+  runAgentRecoveryPhase,
+  type PromptAttemptResult,
+} from "./run-agent-recovery-phase.js";
 import type { AgentTurnCoordinator } from "./turn-coordinator.js";
-import type { AgentOutput, AgentRecoveryDiagnosticEntry, AgentRecoveryMetadata, RetrySettingsProvider, RunAgentOptions } from "./contracts.js";
+import type { AgentOutput, RetrySettingsProvider, RunAgentOptions } from "./contracts.js";
 import { isPendingShutdown } from "../runtime/shutdown-registry.js";
 import {
   beginTrackedPhase,
@@ -88,24 +80,13 @@ export interface RunAgentOrchestratorOptions {
   onError?: (message: string, details: Record<string, unknown>) => void;
 }
 
-async function sleep(ms: number): Promise<void> {
-  if (ms <= 0) return;
-  await Bun.sleep(ms);
-}
-
 const MIN_TOOL_EXECUTION_WATCHDOG_HEARTBEAT_MS = 1_000;
 const MAX_TOOL_EXECUTION_WATCHDOG_HEARTBEAT_MS = 15_000;
 const MID_TURN_CONTEXT_CHECK_MIN_INTERVAL_MS = 1_000;
 const CONTEXT_USAGE_UPDATE_MIN_INTERVAL_MS = 250;
-const MAX_RECOVERY_LOOP_GUARD_CHATS = 512;
 
 /** Approximate characters-per-token ratio for projecting tool-result size onto context tokens. */
 const TOOL_RESULT_CHARS_PER_TOKEN = 4;
-
-type RecoveryFailureSignatureRecord = {
-  atMs: number;
-  signature: string;
-};
 
 type UpstreamAutoCompactionMethod = (...args: unknown[]) => unknown;
 
@@ -114,7 +95,6 @@ type UpstreamAutoCompactionSession = Record<string, unknown> & {
   _runAutoCompaction?: UpstreamAutoCompactionMethod;
 };
 
-const recentRecoveryFailuresByChat = new Map<string, RecoveryFailureSignatureRecord[]>();
 let warnedMissingUpstreamAutoCompactionSuppressorMethods = false;
 
 function suppressUpstreamAutoCompactionDuringPrompt(
@@ -174,87 +154,6 @@ function suppressUpstreamAutoCompactionDuringPrompt(
     if (originalRunAutoCompaction && upstream._runAutoCompaction === runReplacement) {
       upstream._runAutoCompaction = originalRunAutoCompaction;
     }
-  };
-}
-
-function pruneRecoveryFailureMap(now: number, windowMs: number): void {
-  for (const [chatJid, records] of recentRecoveryFailuresByChat.entries()) {
-    const filtered = records.filter((entry) => (now - entry.atMs) <= windowMs);
-    if (filtered.length === 0) {
-      recentRecoveryFailuresByChat.delete(chatJid);
-      continue;
-    }
-    if (filtered.length !== records.length) {
-      recentRecoveryFailuresByChat.set(chatJid, filtered.slice(-200));
-    }
-  }
-
-  if (recentRecoveryFailuresByChat.size <= MAX_RECOVERY_LOOP_GUARD_CHATS) return;
-
-  const oldestChats = Array.from(recentRecoveryFailuresByChat.entries())
-    .map(([chatJid, records]) => ({
-      chatJid,
-      lastAtMs: records.length > 0 ? records[records.length - 1]!.atMs : 0,
-    }))
-    .sort((a, b) => a.lastAtMs - b.lastAtMs);
-
-  const overflow = recentRecoveryFailuresByChat.size - MAX_RECOVERY_LOOP_GUARD_CHATS;
-  for (let i = 0; i < overflow; i += 1) {
-    const candidate = oldestChats[i];
-    if (!candidate) break;
-    recentRecoveryFailuresByChat.delete(candidate.chatJid);
-  }
-}
-
-function normalizeRecoverySignatureText(value: string | null | undefined): string {
-  return String(value || "")
-    .toLowerCase()
-    .replace(/\b\d{2,}\b/g, "#")
-    .replace(/[0-9a-f]{8,}/gi, "#")
-    .replace(/\s+/g, " ")
-    .trim();
-}
-
-function shouldSuppressRecoveryLoop(options: {
-  chatJid: string;
-  modelLabel: string | null;
-  classifier: string;
-  strategy: string | null;
-  errorText: string;
-}): { suppress: boolean; attemptsInWindow: number; maxFailures: number; windowMs: number } {
-  const recoveryPolicy = getRecoveryPolicyConfig();
-  const maxFailures = recoveryPolicy.loopGuardMaxFailures;
-  const windowMs = recoveryPolicy.loopGuardWindowMs;
-  if (!recoveryPolicy.loopGuardEnabled) {
-    return {
-      suppress: false,
-      attemptsInWindow: 0,
-      maxFailures,
-      windowMs,
-    };
-  }
-
-  const now = Date.now();
-  pruneRecoveryFailureMap(now, windowMs);
-
-  const signature = [
-    normalizeRecoverySignatureText(options.modelLabel),
-    normalizeRecoverySignatureText(options.classifier),
-    normalizeRecoverySignatureText(options.strategy),
-    normalizeRecoverySignatureText(options.errorText),
-  ].join("|");
-
-  const current = recentRecoveryFailuresByChat.get(options.chatJid) ?? [];
-  const filtered = current.filter((entry) => (now - entry.atMs) <= windowMs);
-  filtered.push({ atMs: now, signature });
-  recentRecoveryFailuresByChat.set(options.chatJid, filtered.slice(-200));
-
-  const attemptsInWindow = filtered.filter((entry) => entry.signature === signature).length;
-  return {
-    suppress: attemptsInWindow >= maxFailures,
-    attemptsInWindow,
-    maxFailures,
-    windowMs,
   };
 }
 
@@ -453,56 +352,10 @@ function getSessionStateErrorMessage(session: AgentSession): string | null {
   return typeof errorMessage === "string" && errorMessage.trim() ? errorMessage.trim() : null;
 }
 
-interface PromptAttemptResult {
-  output: AgentOutput;
-  snapshot: RecoveryAttemptSnapshot;
-  promptWasPersisted: boolean;
-  timedOut: boolean;
-  toolExecutionCount: number;
-}
-
-function buildRecoveryDiagnosticEntry(
-  phase: AgentRecoveryDiagnosticEntry["phase"],
-  attempt: number,
-  classifier: string,
-  strategy: string | null,
-  reason: string,
-  error: string,
-  elapsedMs: number,
-  snapshot: RecoveryAttemptSnapshot,
-): AgentRecoveryDiagnosticEntry {
-  return {
-    phase,
-    attempt,
-    classifier,
-    strategy,
-    reason,
-    error,
-    elapsedMs,
-    hadToolActivity: Boolean(snapshot.hadToolActivity),
-    hadPartialOutput: Boolean(snapshot.hadPartialOutput),
-    hadCompletedTurnOutput: Boolean(snapshot.hadCompletedTurnOutput),
-    hadTerminalTurnOutput: Boolean(snapshot.hadTerminalTurnOutput),
-    sawCompactionIntent: Boolean(snapshot.sawCompactionIntent),
-    compactionErrorMessage: snapshot.compactionErrorMessage ?? null,
-    toolUseBudgetExceeded: Boolean(snapshot.toolUseBudgetExceeded),
-    assistantToolUseMessageCount: Number.isFinite(snapshot.assistantToolUseMessageCount)
-      ? snapshot.assistantToolUseMessageCount
-      : undefined,
-    toolExecutionCount: Number.isFinite(snapshot.toolExecutionCount)
-      ? snapshot.toolExecutionCount
-      : undefined,
-  };
-}
-
 function resolveToolBudgetWarningThreshold(budget: number): number {
   const normalized = Math.max(1, Math.floor(Number.isFinite(budget) ? budget : 1));
   const margin = Math.min(8, Math.max(1, Math.ceil(normalized * 0.125)));
   return Math.max(1, normalized - margin);
-}
-
-function isAbortFailureText(errorText: string): boolean {
-  return /\b(?:aborterror|aborted|operation was aborted|request was aborted)\b/i.test(errorText);
 }
 
 function buildLengthStopError(partialText: string | null | undefined): string {
@@ -510,80 +363,6 @@ function buildLengthStopError(partialText: string | null | undefined): string {
     ? " The partial answer was preserved; ask me to continue to resume from it."
     : " Ask me to continue and I will retry with a shorter final answer.";
   return `Provider stopped because it hit the maximum output length before finalization (finish reason: length).${partialDetail}`;
-}
-
-function findToolBudgetDiagnostic(diagnostics: AgentRecoveryDiagnosticEntry[]): AgentRecoveryDiagnosticEntry | null {
-  for (let index = diagnostics.length - 1; index >= 0; index -= 1) {
-    const entry = diagnostics[index];
-    const toolExecutionCeilingReached = entry.sawCompactionIntent
-      && Number.isFinite(entry.toolExecutionCount)
-      && (entry.toolExecutionCount ?? 0) > 0;
-    if (entry.toolUseBudgetExceeded
-      || entry.classifier === "tool_history_pressure"
-      || /tool(?:-| )use budget exceeded/i.test(entry.error)
-      || toolExecutionCeilingReached) {
-      return entry;
-    }
-  }
-  return null;
-}
-
-function buildToolBudgetRecoveryTerminalError(
-  budgetDiagnostic: AgentRecoveryDiagnosticEntry,
-  retryErrorText: string,
-): { error: string; toolStepsUsed?: number; toolStepsBudget?: number; nextAction: string } {
-  const parsed = /\((\d+)\/(\d+) tool steps\)/i.exec(budgetDiagnostic.error);
-  const assistantToolUseCount = Number.isFinite(budgetDiagnostic.assistantToolUseMessageCount)
-    ? budgetDiagnostic.assistantToolUseMessageCount
-    : undefined;
-  const toolExecutionCount = Number.isFinite(budgetDiagnostic.toolExecutionCount)
-    ? budgetDiagnostic.toolExecutionCount
-    : undefined;
-  const toolStepsUsed = Number.isFinite(assistantToolUseCount) && (assistantToolUseCount ?? 0) > 0
-    ? assistantToolUseCount
-    : Number.isFinite(toolExecutionCount) && (toolExecutionCount ?? 0) > 0
-      ? toolExecutionCount
-      : parsed ? Number(parsed[1]) : undefined;
-  const toolStepsBudget = parsed ? Number(parsed[2]) : undefined;
-  const budgetDetail = Number.isFinite(toolStepsUsed) && Number.isFinite(toolStepsBudget)
-    ? `Tool-use budget exceeded before finalization (${toolStepsUsed}/${toolStepsBudget} tool steps).`
-    : Number.isFinite(toolStepsUsed)
-      ? `Tool-use budget exceeded before finalization after ${toolStepsUsed} tool execution(s).`
-      : "Tool-use budget exceeded before finalization.";
-  const retryDetail = retryErrorText.trim()
-    ? ` Automatic recovery compacted context and retried, but the retry still produced no terminal assistant reply: ${retryErrorText}`
-    : " Automatic recovery compacted context and retried, but the retry still produced no terminal assistant reply.";
-  const nextAction = "Ask me to continue; I will resume from the latest known partial state instead of replaying the whole turn.";
-  return {
-    error: `${budgetDetail}${retryDetail} ${nextAction}`,
-    toolStepsUsed: Number.isFinite(toolStepsUsed) ? toolStepsUsed : undefined,
-    toolStepsBudget: Number.isFinite(toolStepsBudget) ? toolStepsBudget : undefined,
-    nextAction,
-  };
-}
-
-function buildRecoveryMetadata(
-  attemptsUsed: number,
-  totalElapsedMs: number,
-  recovered: boolean,
-  exhausted: boolean,
-  lastClassifier: string | null,
-  strategyHistory: string[],
-  diagnostics: AgentRecoveryDiagnosticEntry[],
-): AgentRecoveryMetadata {
-  return {
-    attemptsUsed,
-    totalElapsedMs,
-    recovered,
-    exhausted,
-    lastClassifier,
-    strategyHistory,
-    diagnostics,
-  };
-}
-
-function emitAgentSessionEvent(onEvent: RunAgentOptions["onEvent"], event: Record<string, unknown>): void {
-  onEvent?.(event as AgentSessionEvent);
 }
 
 function estimatePendingInputTokens(prompt: string): number {
@@ -613,101 +392,6 @@ function getRunObservabilityDetails(
     ...(runOptions.clientId ? { clientId: runOptions.clientId } : {}),
     ...(sessionLeafId ? { sessionLeafId } : {}),
   };
-}
-
-async function runRecoveryCompaction(
-  session: AgentSession,
-  chatJid: string,
-  runOptions: RunAgentOptions,
-  options: Pick<RunAgentOrchestratorOptions, "onInfo" | "onWarn">,
-): Promise<{ ok: true } | { ok: false; errorMessage: string }> {
-  options.onInfo?.("Compacting before automatic recovery retry", {
-    operation: "run_agent.recovery_compact",
-    chatJid,
-  });
-  const retryMetadata: PiclawCompactionTriggerMetadata = {
-    chatJid,
-    trigger: "recovery",
-    willRetry: true,
-    source: "automatic_recovery",
-  };
-  const retryEventFields = buildPiclawCompactionEventFields(retryMetadata, { reason: "overflow", willRetry: true });
-  emitAgentSessionEvent(runOptions.onEvent, { type: "compaction_start", ...retryEventFields } as unknown as AgentSessionEvent);
-  heartbeatTrackedPhase(chatJid, "preprompt_compaction", {
-    eventType: "recovery_compaction_start",
-    source: "automatic_recovery",
-    ...getRunObservabilityDetails(runOptions),
-  });
-  const compactionResult = await runCompactionWithTimeout(
-    session,
-    chatJid,
-    options,
-    async () => await session.compact(),
-    "recovery",
-    { trigger: "recovery", willRetry: true, source: "automatic_recovery" },
-  );
-  heartbeatTrackedPhase(chatJid, "recovery", {
-    eventType: "recovery_compaction_end",
-    ok: compactionResult.ok,
-    ...getRunObservabilityDetails(runOptions),
-  });
-  if (!compactionResult.ok) {
-    const aborted = isCompactionCancellationError(compactionResult.errorMessage);
-    const benign = isRotationFallbackCompactionError(compactionResult.errorMessage);
-    if (!compactionResult.joined && !aborted && !benign) {
-      noteCompactionFailure(chatJid, compactionResult.errorMessage);
-    }
-    if (benign) {
-      // "Nothing to compact (session too small)" etc. — not a real failure,
-      // skip compaction and let the retry proceed without it.
-      options.onInfo?.("Recovery compaction skipped (benign: session too small or already compacted)", {
-        operation: "run_agent.recovery_compact_benign_skip",
-        chatJid,
-        errorMessage: compactionResult.errorMessage,
-      });
-      emitAgentSessionEvent(runOptions.onEvent, {
-        type: "compaction_end",
-        ...retryEventFields,
-        result: undefined,
-        aborted: false,
-        willRetry: true,
-      } as unknown as AgentSessionEvent);
-      return { ok: true };
-    }
-    emitAgentSessionEvent(runOptions.onEvent, {
-      type: "compaction_end",
-      ...buildPiclawCompactionEventFields(retryMetadata, { reason: "overflow", willRetry: false }),
-      result: undefined,
-      aborted,
-      willRetry: false,
-      errorMessage: aborted ? undefined : `Recovery compaction failed: ${compactionResult.errorMessage}`,
-    } as unknown as AgentSessionEvent);
-    return { ok: false, errorMessage: compactionResult.errorMessage };
-  }
-  if (!compactionResult.joined) {
-    noteCompactionSuccess(session, chatJid, "recovery", {
-      ...options,
-      countSuccess: false,
-    });
-  }
-  const contextReport = getCompactionContextReport(session, compactionResult.result as { tokensBefore?: unknown; estimatedTokensAfter?: unknown });
-  emitAgentSessionEvent(runOptions.onEvent, {
-    type: "compaction_end",
-    ...retryEventFields,
-    result: compactionResult.result,
-    aborted: false,
-    willRetry: true,
-    tokensBefore: contextReport.tokensBefore ?? undefined,
-    estimatedTokensAfter: contextReport.estimatedTokensAfter,
-    estimatedTokensAfterSource: contextReport.estimatedTokensAfterSource,
-    safetyAdjustedTokensAfter: contextReport.safetyAdjustedTokensAfter,
-    reductionPercent: contextReport.reductionPercent ?? undefined,
-  } as unknown as AgentSessionEvent);
-  const usageEvent = buildFreshContextUsageUpdateEvent(session, chatJid, "after_recovery_compaction", {
-    source: "compaction_recovery",
-  });
-  if (usageEvent) emitAgentSessionEvent(runOptions.onEvent, usageEvent);
-  return { ok: true };
 }
 
 async function runPromptAttempt(
@@ -1729,428 +1413,34 @@ export async function runAgentPrompt(
     const recoveryConfig = timeoutMs > 0
       ? { ...baseRecoveryConfig, totalBudgetMs: Math.min(baseRecoveryConfig.totalBudgetMs, timeoutMs) }
       : baseRecoveryConfig;
-    let recoveryAttemptsUsed = 0;
-    let lastClassifier: RecoveryClassifier | null = null;
-    let lastRecoveryErrorText: string | null = null;
-    const strategyHistory: RecoveryStrategy[] = [];
-    const recoveryDiagnostics: AgentRecoveryDiagnosticEntry[] = [];
-    let recoveryBudgetStartedAt: number | null = null;
-    let recoveryBudgetAccumulatedMs = 0;
-    let allowPostTimeoutRecoveryWindow = false;
-    const getRecoveryBudgetElapsedMs = () => {
-      if (recoveryBudgetStartedAt != null) {
-        return Math.max(0, recoveryBudgetAccumulatedMs + Date.now() - recoveryBudgetStartedAt);
-      }
-      if (recoveryBudgetAccumulatedMs > 0) return recoveryBudgetAccumulatedMs;
-      return timeoutMs <= 0 || allowPostTimeoutRecoveryWindow
-        ? 0
-        : Math.max(0, Date.now() - startTime);
-    };
-    const getRecoveryDecisionElapsedMs = (errorText: string, snapshot: RecoveryAttemptSnapshot) => {
-      // #778: context-pressure compact_then_retry has two bounded phases:
-      // recovery compaction is bounded by the compaction timeout; the
-      // continuation prompt receives a fresh recovery budget after compaction.
-      // Therefore the initial prompt duration must not exhaust the recovery
-      // decision before compaction can run.
-      if (recoveryAttemptsUsed === 0 && (isContextPressureFailure(errorText) || snapshot.sawCompactionIntent)) return 0;
-      return getRecoveryBudgetElapsedMs();
-    };
-    const startRecoveryBudget = () => {
-      if (recoveryBudgetStartedAt == null) recoveryBudgetStartedAt = Date.now();
-    };
-    const pauseRecoveryBudget = () => {
-      if (recoveryBudgetStartedAt == null) return;
-      recoveryBudgetAccumulatedMs += Math.max(0, Date.now() - recoveryBudgetStartedAt);
-      recoveryBudgetStartedAt = null;
-    };
 
-    const runResult: AgentOutput = await withChatContext(chatJid, channel, async () => {
-      let attemptPrompt = prompt;
-      let recoveryContinuationWithoutTools = false;
-      let turnToolExecutionCount = 0;
-      while (true) {
-        // Yield to the event loop on every iteration. Prevents synchronous-
-        // throw + catch + retry from starving the event loop when the error
-        // path never reaches an await that actually suspends.
-        await new Promise<void>((resolve) => setTimeout(resolve, 0));
-
-        if (recoveryAttemptsUsed > 0 && getRecoveryBudgetElapsedMs() >= recoveryConfig.totalBudgetMs) {
-          const duration = Date.now() - startTime;
-          const error = lastRecoveryErrorText || "Automatic recovery budget exhausted before the next attempt could start.";
-          lastClassifier = "budget_exhausted";
-          const recovery = buildRecoveryMetadata(
-            recoveryAttemptsUsed,
-            duration,
-            false,
-            true,
-            lastClassifier,
-            strategyHistory,
-            recoveryDiagnostics,
-          );
-          writeAgentLog(options.logsDir, chatJid, duration, false, null, error, recovery);
-          emitAgentSessionEvent(runOptions.onEvent, {
-            type: "recovery_end",
-            outcome: "exhausted",
-            attemptsUsed: recoveryAttemptsUsed,
-            classifier: lastClassifier,
-            errorMessage: error,
-          });
-          return { status: "error" as const, result: null, error, recovery };
-        }
-
-        // The configured turn timeout bounds the initial attempt. Once that
-        // attempt fails after tool activity, automatic recovery gets its own
-        // smaller total budget; otherwise a timeout failure can never start
-        // its safe tools-disabled continuation.
-        if (timeoutMs > 0 && recoveryAttemptsUsed === 0) {
-          const loopElapsedMs = Date.now() - startTime;
-          if (loopElapsedMs > timeoutMs) {
-            const duration = Date.now() - startTime;
-            writeAgentLog(options.logsDir, chatJid, duration, false, null,
-              `Agent run exceeded timeout before recovery (${loopElapsedMs}ms > ${timeoutMs}ms)`);
-            return {
-              status: "error" as const,
-              result: null,
-              error: `Agent run timed out after ${Math.round(loopElapsedMs / 1000)}s`,
-            };
-          }
-        }
-
-        const remainingRecoveryBudgetMs = Math.max(1, recoveryConfig.totalBudgetMs - getRecoveryBudgetElapsedMs());
-        const attemptTimeoutMs = recoveryAttemptsUsed > 0
-          ? (timeoutMs > 0 ? Math.min(timeoutMs, remainingRecoveryBudgetMs) : remainingRecoveryBudgetMs)
-          : timeoutMs;
-        let recoverySavedToolNames: string[] | null = null;
-        if (recoveryContinuationWithoutTools && sessionCtrl && typeof sessionCtrl.getActiveToolNames === "function" && typeof sessionCtrl.setActiveToolsByName === "function") {
-          recoverySavedToolNames = sessionCtrl.getActiveToolNames();
-          sessionCtrl.setActiveToolsByName([]);
-        }
-        let attempt: PromptAttemptResult;
-        try {
-          attempt = await runPromptAttempt(
-            attemptPrompt,
-            chatJid,
-            session,
-            attemptTimeoutMs,
-            runOptions,
-            options,
-            startTime,
-            modelLabel,
-            turnToolExecutionCount,
-          );
-          turnToolExecutionCount = attempt.toolExecutionCount;
-        } finally {
-          if (recoverySavedToolNames && sessionCtrl && typeof sessionCtrl.setActiveToolsByName === "function") {
-            sessionCtrl.setActiveToolsByName(recoverySavedToolNames);
-          }
-        }
-
-        // If the tool-call cap was hit, abort immediately without recovery.
-        if (toolCallCapRef.exceeded) {
-          const duration = Date.now() - startTime;
-          const used = Number.isFinite(toolCallCapRef.count) ? toolCallCapRef.count : undefined;
-          const cap = Number.isFinite(toolCallCapRef.cap) ? toolCallCapRef.cap : undefined;
-          const budgetDetail = Number.isFinite(used) && Number.isFinite(cap)
-            ? `Tool-use budget exceeded before finalization (${used}/${cap} tool calls).`
-            : Number.isFinite(used)
-              ? `Tool-use budget exceeded before finalization after ${used} tool call(s).`
-              : "Tool-use budget exceeded before finalization.";
-          const nextAction = "Ask me to continue; I will resume from the latest known partial state instead of replaying the whole turn.";
-          const error = `${budgetDetail} ${nextAction}`;
-          writeAgentLog(options.logsDir, chatJid, duration, false, null, error);
-          return {
-            status: "error",
-            result: null,
-            error,
-            toolBudgetExceeded: true,
-            toolStepsUsed: used,
-            toolStepsBudget: cap,
-            nextAction,
-          };
-        }
-
-        if (attempt.output.status === "success") {
-          const duration = Date.now() - startTime;
-          const finalText = typeof attempt.output.result === "string" ? attempt.output.result : null;
-          const recoveryMeta = recoveryAttemptsUsed > 0
-            ? buildRecoveryMetadata(recoveryAttemptsUsed, duration, true, false, lastClassifier, strategyHistory, recoveryDiagnostics)
-            : null;
-          writeAgentLog(options.logsDir, chatJid, duration, false, finalText, null, recoveryMeta);
-          options.onInfo?.("Agent run completed", {
-            operation: "run_agent.complete",
-            chatJid,
-            model: modelLabel,
-            durationMs: duration,
-            outputChars: finalText?.length ?? 0,
-            recoveryAttemptsUsed,
-            recovered: recoveryAttemptsUsed > 0,
-            ...getRunObservabilityDetails(runOptions),
-          });
-          if (recoveryAttemptsUsed > 0) {
-            emitAgentSessionEvent(runOptions.onEvent, {
-              type: "recovery_end",
-              outcome: "recovered",
-              attemptsUsed: recoveryAttemptsUsed,
-              classifier: lastClassifier,
-            });
-            attempt.output.recovery = buildRecoveryMetadata(
-              recoveryAttemptsUsed,
-              duration,
-              true,
-              false,
-              lastClassifier,
-              strategyHistory,
-              recoveryDiagnostics,
-            );
-          }
-          recentRecoveryFailuresByChat.delete(chatJid);
-          return attempt.output;
-        }
-
-        if (attempt.output.status === "tool_complete") {
-          const duration = Date.now() - startTime;
-          writeAgentLog(options.logsDir, chatJid, duration, false, null, null, null);
-          options.onInfo?.("Agent run completed via terminal tool", {
-            operation: "run_agent.tool_complete",
-            chatJid,
-            model: modelLabel,
-            durationMs: duration,
-            ...getRunObservabilityDetails(runOptions),
-          });
-          recentRecoveryFailuresByChat.delete(chatJid);
-          return attempt.output;
-        }
-
-        const errorText = attempt.output.error || "Agent error";
-        lastRecoveryErrorText = errorText;
-        allowPostTimeoutRecoveryWindow = recoveryAttemptsUsed === 0
-          && attempt.snapshot.hadToolActivity
-          && attempt.timedOut;
-        const decision = decideAutomaticRecovery({
-          config: recoveryConfig,
-          errorText,
-          recoveryAttemptsUsed,
-          elapsedMs: getRecoveryDecisionElapsedMs(errorText, attempt.snapshot),
-          snapshot: attempt.snapshot,
-        });
-
-        let effectiveDecision = decision;
-        if (decision.recover && decision.strategy) {
-          const guard = shouldSuppressRecoveryLoop({
-            chatJid,
-            modelLabel,
-            classifier: decision.classifier,
-            strategy: decision.strategy,
-            errorText,
-          });
-          if (guard.suppress) {
-            effectiveDecision = {
-              recover: false,
-              classifier: "recovery_suppressed",
-              strategy: null,
-              reason: `Automatic recovery suppressed after ${guard.attemptsInWindow} repeated failures within ${Math.max(1, Math.round(guard.windowMs / 60000))} minute(s).`,
-            };
-          }
-        }
-
-        lastClassifier = effectiveDecision.classifier;
-
-        options.onWarn?.("Agent attempt failed", {
-          operation: "run_agent.attempt_failed",
-          chatJid,
-          errorText,
-          classifier: effectiveDecision.classifier,
-          recoveryAttemptsUsed,
-          recoveryStrategy: effectiveDecision.strategy,
-          reason: effectiveDecision.reason,
-          ...getRunObservabilityDetails(runOptions),
-        });
-
-        recoveryDiagnostics.push(buildRecoveryDiagnosticEntry(
-          "attempt_failure",
-          recoveryAttemptsUsed + 1,
-          effectiveDecision.classifier,
-          effectiveDecision.strategy,
-          effectiveDecision.reason,
-          errorText,
-          Date.now() - startTime,
-          attempt.snapshot,
-        ));
-
-        if (!effectiveDecision.recover || !effectiveDecision.strategy) {
-          const duration = Date.now() - startTime;
-          const toolBudgetDiagnostic = recoveryAttemptsUsed > 0 ? findToolBudgetDiagnostic(recoveryDiagnostics) : null;
-          const terminalBudgetFailure = toolBudgetDiagnostic
-            && (/Prompt completed without emitting an assistant reply before finalization/i.test(errorText) || isAbortFailureText(errorText))
-            ? buildToolBudgetRecoveryTerminalError(toolBudgetDiagnostic, errorText)
-            : null;
-          const finalErrorText = terminalBudgetFailure?.error ?? errorText;
-          const finalClassifier = terminalBudgetFailure && toolBudgetDiagnostic ? toolBudgetDiagnostic.classifier : lastClassifier;
-          const recoveryMeta = (recoveryAttemptsUsed > 0 || Boolean(finalClassifier))
-            ? buildRecoveryMetadata(recoveryAttemptsUsed, duration, false, true, finalClassifier, strategyHistory, recoveryDiagnostics)
-            : null;
-          writeAgentLog(options.logsDir, chatJid, duration, false, null, finalErrorText, recoveryMeta);
-          if (recoveryAttemptsUsed > 0 || effectiveDecision.classifier === "recovery_suppressed") {
-            emitAgentSessionEvent(runOptions.onEvent, {
-              type: "recovery_end",
-              outcome: "exhausted",
-              attemptsUsed: recoveryAttemptsUsed,
-              classifier: finalClassifier ?? effectiveDecision.classifier,
-              errorMessage: finalErrorText,
-            });
-          }
-          if (recoveryMeta) {
-            attempt.output.recovery = recoveryMeta;
-          }
-          if (terminalBudgetFailure) {
-            attempt.output.error = finalErrorText;
-            attempt.output.toolBudgetExceeded = true;
-            attempt.output.toolStepsUsed = terminalBudgetFailure.toolStepsUsed;
-            attempt.output.toolStepsBudget = terminalBudgetFailure.toolStepsBudget;
-            attempt.output.nextAction = terminalBudgetFailure.nextAction;
-          }
-          return attempt.output;
-        }
-
-        recoveryAttemptsUsed += 1;
-        strategyHistory.push(effectiveDecision.strategy);
-        const retryDelayMs = effectiveDecision.strategy === "retry"
-          ? getAutomaticRecoveryDelayMs(recoveryConfig, recoveryAttemptsUsed)
-          : 0;
-        heartbeatTrackedPhase(chatJid, "recovery", {
-          eventType: "recovery_start",
-          attempt: recoveryAttemptsUsed,
-        });
-        emitAgentSessionEvent(runOptions.onEvent, {
-          type: "recovery_start",
-          classifier: effectiveDecision.classifier,
-          strategy: effectiveDecision.strategy,
-          attempt: recoveryAttemptsUsed,
-          maxAttempts: recoveryConfig.maxAttempts,
-          totalBudgetMs: recoveryConfig.totalBudgetMs,
-          delayMs: retryDelayMs,
-          reason: effectiveDecision.classifier === "unknown" && errorText
-            ? `${effectiveDecision.reason} Error: ${errorText}`
-            : effectiveDecision.reason,
-          errorMessage: errorText,
-        });
-
-        if (effectiveDecision.strategy !== "compact_then_retry"
-          && recoveryBudgetStartedAt == null
-          && recoveryBudgetAccumulatedMs === 0
-          && (timeoutMs <= 0 || allowPostTimeoutRecoveryWindow)) {
-          startRecoveryBudget();
-        }
-
-        if (retryDelayMs > 0) {
-          heartbeatTrackedPhase(chatJid, "recovery", {
-            eventType: "recovery_delay",
-            delayMs: retryDelayMs,
-          });
-          await sleep(retryDelayMs);
-        }
-
-        // AgentSession persists the user message before invoking the provider.
-        // Replaying the original text after a failed attempt duplicates the
-        // instruction and can repeat side-effecting tools. Resume persisted
-        // turns with a neutral continuation; only replay when no branch state
-        // was appended (for example, a synchronous pre-prompt throw).
-        if (attempt.promptWasPersisted || attempt.snapshot.hadToolActivity) {
-          attemptPrompt = RECOVERY_CONTINUATION_PROMPT;
-        }
-        recoveryContinuationWithoutTools = recoveryContinuationWithoutTools || attempt.snapshot.hadToolActivity;
-
-        if (effectiveDecision.strategy === "compact_then_retry") {
-          pauseRecoveryBudget();
-          const compactionResult = await runRecoveryCompaction(session, chatJid, runOptions, options);
-          heartbeatTrackedPhase(chatJid, "preprompt_compaction", {
-            eventType: "recovery_compaction",
-            attempt: recoveryAttemptsUsed,
-          });
-          if (!compactionResult.ok) {
-            const compactDecision = decideAutomaticRecovery({
-              config: recoveryConfig,
-              errorText: compactionResult.errorMessage,
-              recoveryAttemptsUsed,
-              elapsedMs: getRecoveryBudgetElapsedMs(),
-              snapshot: {
-                hadToolActivity: false,
-                hadPartialOutput: attempt.snapshot.hadPartialOutput,
-                hadCompletedTurnOutput: attempt.snapshot.hadCompletedTurnOutput,
-                hadTerminalTurnOutput: attempt.snapshot.hadTerminalTurnOutput,
-                compactionErrorMessage: compactionResult.errorMessage,
-                sawCompactionIntent: true,
-                toolUseBudgetExceeded: attempt.snapshot.toolUseBudgetExceeded,
-                assistantToolUseMessageCount: attempt.snapshot.assistantToolUseMessageCount,
-                toolExecutionCount: attempt.snapshot.toolExecutionCount,
-              },
-            });
-            lastClassifier = compactDecision.classifier;
-            if (!compactDecision.recover || compactDecision.strategy !== "retry") {
-              recoveryDiagnostics.push(buildRecoveryDiagnosticEntry(
-                "compaction_failure",
-                recoveryAttemptsUsed,
-                compactDecision.classifier,
-                compactDecision.strategy,
-                compactDecision.reason,
-                compactionResult.errorMessage,
-                Date.now() - startTime,
-                {
-                  hadToolActivity: false,
-                  hadPartialOutput: attempt.snapshot.hadPartialOutput,
-                  hadCompletedTurnOutput: attempt.snapshot.hadCompletedTurnOutput,
-                  hadTerminalTurnOutput: attempt.snapshot.hadTerminalTurnOutput,
-                  compactionErrorMessage: compactionResult.errorMessage,
-                  sawCompactionIntent: true,
-                  toolUseBudgetExceeded: attempt.snapshot.toolUseBudgetExceeded,
-                  assistantToolUseMessageCount: attempt.snapshot.assistantToolUseMessageCount,
-                  toolExecutionCount: attempt.snapshot.toolExecutionCount,
-                },
-              ));
-              const duration = Date.now() - startTime;
-              const recoveryMeta = buildRecoveryMetadata(
-                recoveryAttemptsUsed,
-                duration,
-                false,
-                true,
-                lastClassifier,
-                strategyHistory,
-                recoveryDiagnostics,
-              );
-              writeAgentLog(options.logsDir, chatJid, duration, false, null, compactionResult.errorMessage, recoveryMeta);
-              emitAgentSessionEvent(runOptions.onEvent, {
-                type: "recovery_end",
-                outcome: "exhausted",
-                attemptsUsed: recoveryAttemptsUsed,
-                classifier: compactDecision.classifier,
-                errorMessage: compactionResult.errorMessage,
-              });
-              return {
-                status: "error",
-                result: null,
-                error: compactionResult.errorMessage,
-                recovery: buildRecoveryMetadata(
-                  recoveryAttemptsUsed,
-                  duration,
-                  false,
-                  true,
-                  lastClassifier,
-                  strategyHistory,
-                  recoveryDiagnostics,
-                ),
-              };
-            }
-          }
-          startRecoveryBudget();
-        }
-
-        heartbeatTrackedPhase(chatJid, "prompt", {
-          eventType: "recovery_retry_ready",
-          attempt: recoveryAttemptsUsed,
-        });
-        options.clearAttachments(chatJid);
-      }
-    });
+    const runResult: AgentOutput = await withChatContext(chatJid, channel, async () => await runAgentRecoveryPhase({
+      prompt,
+      chatJid,
+      session,
+      sessionCtrl,
+      timeoutMs,
+      startTime,
+      modelLabel,
+      recoveryConfig,
+      runOptions,
+      logsDir: options.logsDir,
+      onInfo: options.onInfo,
+      onWarn: options.onWarn,
+      clearAttachments: options.clearAttachments,
+      toolCallCap: toolCallCapRef,
+      runPromptAttempt: async (attemptPrompt, attemptTimeoutMs, turnToolExecutionCount) => await runPromptAttempt(
+        attemptPrompt,
+        chatJid,
+        session,
+        attemptTimeoutMs,
+        runOptions,
+        options,
+        startTime,
+        modelLabel,
+        turnToolExecutionCount,
+      ),
+    }));
 
     if (runOptions.scheduleIdleAutoCompaction && (runResult.status === "success" || runResult.status === "tool_complete")) {
       await maybeAutoCompactSessionAfterTurn(session, chatJid, options, (event) => {
